[broken][skip ci] Fix the aggregation of variables and profiles
In the incoming and outgoing expressions, use an aggregating function
to decide how to aggregate. Either simply sum the variables (which
results in the duration being multiplied), or apply the unique functions
before (which removes the duration from the expression).

In the profiles, change the profile_sum to profile_aggregation and
require a function to be passed. This was only the sum, but it can be
the average now.