Implement all different resolutions (#469)
* [broken][skip ci] Implement different constraints partitions

There are 5 constraints partitions now, used for different constraints,
depending on the type of flow:

- power - use the :highest resolution
- energy - use the :lowest resolution

And also on what are the input partitions:

- partitions of incoming flows
- partitions of outgoing flows
- partitions of the asset itself

* [broken][skip ci] Add Statistics and OrderedDict

* [broken][skip ci] Update construct_dataframes to use the new constraints partitions

The main change is that there is now a dedicated
`constraints_partitions[:lowest_storage_balance]` for the balance of
storage assets. This means that we can use a single dataframe for those
constraints and for the storage level variable.

* [broken][skip ci] Fix the aggregation of variables and profiles

In the incoming and outgoing expressions, use an aggregating function
to decide how to aggregate. Either simply sum the variables (which
results in the duration being multiplied), or apply the unique functions
before (which removes the duration from the expression).

In the profiles, change the profile_sum to profile_aggregation and
require a function to be passed. This was only the sum, but it can be
the average now.

* [broken][skip ci] Change the model to use the last partition changes

Try to follow #344 and implement the correct usage of the dataframes and
the aggregation.

* [broken][skip ci] Add a name to the max_storage_level_limit constraint

* [broken][skip ci] Fix conversion asset using wrong resolution

* Rename :lowest_storage_balance to :lowest_storage_level

* [broken][skip ci] Fix some things that were not fixed in the last two commits

* [skip ci] Fix tests

* [broken] Try to fix documentation

* Move profile_aggregation to the outside of the create_model function

* Delete test/Manifest.toml and update .gitignore

Change /Manifest.toml and /docs/Manifest.toml to Manifest.toml so
it catches all cases.

* Uncomment drop_zeros!

---------

Co-authored-by: Abel Soares Siqueira <abel.siqueira@esciencecenter.nl>