Use Symbols instead of Strings for many cases
Most notable cases are the names and types of assets and profiles.
Some loose keys were also included in this. Note that partition
definitions are still Strings, since they will be manipulated as
Strings.
The only file change was the name of the max and min profiles,
since they had dashes in their name. To avoid changing the file name,
I sanitize it to convert - to _.