Create constraints partitions as DataFrames
Creates compute_constraints_partitions functions to use TableTree to
build the partitions of the constraints. This is done in two parts.

First, we create unrolled partitions, which will substitute the
individual partitions vectors stores in the graphs.

Second, we build constraints partitions still using
compute_rp_partitions, but using only DataFrames. The result is also
stored in a DataFrame, which should also serve as the final DataFrame to
be used instead of energy_problem.dataframes.