Fix CompareMPS when files differ in length
If the files had different length, the script prints errors, but was not
returning errors. This returns the correct exit flag.

We also update the message to make it easier to find the comment and try
to avoid duplicate CompareMPS messages.

Finally, we clarify that forks won't have the messages.

Closes #1188