Rolling horizon (#1327)
* Change rep_period profiles to handle rolling horizon

Implement ProfileWithRollingHorizon and make profiles.rep_period use
that instead of Vector{Float64}.
Change _profile_aggregate to dispatch on Vector{Float64} and
ProfileWithRollingHorizon.
These should allow the rolling horizon variables to be aggregated
the same way as the normal profiles.

Part of #1365

* Implement high-level run_rolling_horizon function with placeholders

Implement run_rolling_horizon with the main idea. Use placeholders
to specify places where a more complicated function needs to be called.
Update the EnergyProblem structure to hold an inner EnergyProblem
for the rolling horizon model.

Part of #1365

* Create rolling horizon parameters

Add ParametricOptInterface.
Implement adding the rolling horizon parameters in src/rolling-horizon/create.jl.
Update create_model to handle the rolling horizon parameters and wrap the
solver in POI.Optimizer if it's rolling horizon.

Part of #1365

* Implement update of rolling horizon parameters

Implement the update functions for rolling horizon and add them
to the run_rolling_horizon function.

Part of #1365

* Test adding rolling horizon parameters

Part of #1365

* Add functions for validation and data preparation for rolling horizon

Add validation of input in the rolling horizon function.
Add function to prepare the input to get in the rolling horizon and to get
out.
Add function to save the solution in the relevant tables.

Part of #1365

* Add unit tests for rolling horizon

Add tests of the various parts of rolling horizon.

Part of #1365

* Update the storage constraint to use the rolling initial storage level

Update src/constraints/storage.jl to use the initial_storage_level stored
in the rolling horizon.

Add test for the rolling horizon objective values to control future changes.

Part of #1365