Compute and save dual variables during rolling horizon (#1407)
Add an argument compute_duals to run_rolling_horizon that is passed to
save_solution.
Save the cons_% tables as full_cons_% during rolling horizon.
If the duals are computed, create a column dual_% in the full_cons_%
table and copy the internal dual_% value from the cons_% table in the
window to the corresponding full_cons_% table.

Closes #1395